-New --Added Classes for new tabbed pane. --this will when complete will split the user screen and admin screen class into much smaller manageable classes --Create booking is 70% complete Positioning needs to finished --Layouts and components will all need to be set into new layouts as when split up the null/absolute layout from the UI builder is not working correctly anymore.
N.B I have not altered the main program yet so anything your running or testing u will still use the main method in the start screen class.
I wont change this until all screens are completed.

This is going to take most of the weekend to do.
tinkered with using the model more than we currently are. on a local copy hoping to show an alternative approach on monday the 24th.